requirements.txt
flask==2.3.3
google-generativeai==0.3.1
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.3.0
